added pocketbase config. moved tensorflow to new file (tf_utils.py) and combined app.py and main.py. deleted files no longer needed or used for testing.
