--- index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-    <form action="/" method="post" enctype="multipart/form-data">
-        <input type="file" name="file" accept="image/*">
-        <button type="submit">Upload</button>
-    </form>
-
-    {% if filename %}
-        <img src="{{ url_for('static', filename='uploads/' + filename) }}" width="300">
-        <p>{{ label }} ({{ confidence|round(2) }}%)</p>
-    {% endif %}
-</body>
-</html>
