+++ templates/diagnose_plant.html
@@ -1,13 +1,14 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>plant</title>
-    <link href="/templates/style.css" rel="stylesheet" type="text/css" media="all" />
-  </head>
+<body>
+    <form action="/" method="post" enctype="multipart/form-data">
+        <input type="file" name="file" accept="image/*">
+        <button type="submit">Upload</button>
+    </form>
 
-  <body>
-    <input type="file" id="file" name="avatar"/>
-  </body>
+    {% if filename %}
+        <img src="{{ url_for('static', filename='uploads/' + filename) }}" width="300">
+        <p>{{ label }} ({{ confidence|round(2) }}%)</p>
+    {% endif %}
+</body>
 </html>
